refactor contact components to update import paths for useContactActions; remove index file from hooks
