--- src/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './use-contact-actions';
-export * from './use-contacts-table';
-export * from './use-mobile';
